pyproject.toml
[project]
name = "llm-interface"
version = "0.1.0"
description = "A unified interface for querying multiple LLM models"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
requires-python = ">=3.8"
dependencies = [
    "openai>=1.12.0",
    "transformers>=4.37.0",
    "torch>=2.2.0",
    "huggingface-hub>=0.20.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "black>=24.1.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=llm_interface --cov-report=term-missing" 